==> 01_MetRS_CS_AnalyticFile.Rmd <==
---
title: "FEV1_metRS_CS"
author: "Suneeta Godbole"
date: "7/14/2021"
output: pdf_document
---

```{r setup}
rm(list=ls())
library(glmnet)

library(ggplot2)
library(grid)
library(gridExtra)
library(gridtext)
library(VennDiagram)

datafiles <- "C:/NetCo/Metabolomic_Risk_Score/metRS_CS_AnalyticFiles"
spiro.metInfo.folder <- "C:/NetCo/Metabolomics/NewMetabolonDatasets_20211018/processed"

## Data Read-In
spiro.c <- readRDS(file.path(datafiles, "SPIROMICS_MetRS_CS_Clinical.rds"))
met_all <- readRDS(file.path(datafiles, "SPIROMICS_MetRS_CS_Metabolites.rds"))
var.key <- readRDS(file.path(spiro.metInfo.folder, "SPIROMICS_V1_MetaboliteInformation_20211021.rds"))

copd.clin.p2 <- readRDS(file.path(datafiles, "COPDGene_P2_MetRS_CS_Clinical.rds"))
metRS_copd <- readRDS(file.path(datafiles, "COPDGene_P2_MetRS_CS_Metabolites.rds"))

## Data outfile
metRSfolder <- "C:/NetCo/Metabolomic_Risk_Score/metRS_output"

## Function to caluculate MSE  
fxn.mse <- function(x, y){
  mse <- mean((x-y)^2)
  return(mse)
}

```

# FEV1

## Data prep
```{r fev1_dataprep}
# SPIROMICS -- data prep
met_all$SUBJID <- rownames(met_all)
spiro.fev1 <- merge(spiro.c[, c("SUBJID", "POST_FEV1_DERV_V1")], 
                       met_all, by = "SUBJID")

spiro.fev1 <- spiro.fev1[!(is.na(spiro.fev1$POST_FEV1_DERV_V1)), ] # Remove subj with missing FEV1

rownames(spiro.fev1) <- spiro.fev1$SUBJID
spiro.fev1$SUBJID <- NULL

# COPDGene -- data prep
metRS_copd$sid <- rownames(metRS_copd)

copdg.fev1 <- merge(copd.clin.p2[, c("sid", "FEV1_post.2")], 
                       metRS_copd, by = "sid")
copdg.fev1 <- copdg.fev1[!(is.na(copdg.fev1$FEV1_post.2)), ]

rownames(copdg.fev1) <- copdg.fev1$sid
copdg.fev1$sid <-  NULL
```

## FEV1 Adaptive LASSO Results Tables
```{r fev1_adaLASSO}
fev1.adaLASSO.tS <- data.frame("study" = c("SPIROMICS", "COPDGene"), 
                            stringsAsFactors = F)

fev1.adaLASSO.tC <- data.frame("study" = c("SPIROMICS", "COPDGene"), 
                            stringsAsFactors = F)
fev1.adaLASSO.tSwt <- data.frame("study" = c("SPIROMICS", "COPDGene"), 
                            stringsAsFactors = F)
```

## Adjust R-squared value for covariate only models
```{r covarOnly}
spiro.covar <- lm(POST_FEV1_DERV_V1 ~ GENDER + AGE_DERV_V1 + RACE.ETH + 
                      BMI_CM_V1 + smokStatus + SMOKING_PACK_YEARS_V1 + SITE_cat,
                      data = spiro.c)
spiro.covarOnly <- round(summary(spiro.covar)$adj.r.squared*100, 3) #35.391

copdg.covar <- lm(FEV1_post.2 ~ gender + age_visit + race.eth + 
                    BMI + smoking_status + ATS_PackYears + ccenter, 
                    data = copd.clin.p2)
copdg.covarOnly <- round(summary(copdg.covar)$adj.r.squared*100, 3) #40.213

fev1.adaLASSO.tS$Cov_Only_R2 <- c(spiro.covarOnly, copdg.covarOnly)
fev1.adaLASSO.tS$Cov_Only_MSE <- c(fxn.mse(spiro.covar$model$POST_FEV1_DERV_V1, fitted(spiro.covar)), 
                                fxn.mse(copdg.covar$model$FEV1_post.2, fitted(copdg.covar)))

fev1.adaLASSO.tC$Cov_Only_R2 <- c(spiro.covarOnly, copdg.covarOnly)
fev1.adaLASSO.tC$Cov_Only_MSE <- c(fxn.mse(spiro.covar$model$POST_FEV1_DERV_V1, fitted(spiro.covar)), 
                                fxn.mse(copdg.covar$model$FEV1_post.2, fitted(copdg.covar)))

fev1.adaLASSO.tSwt$Cov_Only_R2 <- c(spiro.covarOnly, copdg.covarOnly)
fev1.adaLASSO.tSwt$Cov_Only_MSE <- c(fxn.mse(spiro.covar$model$POST_FEV1_DERV_V1, fitted(spiro.covar)), 
                                fxn.mse(copdg.covar$model$FEV1_post.2, fitted(copdg.covar)))
``` 

## ADAPTIVE LASSO -- Train SPIROMICS
```{r adalasso_spiro}
set.seed(20210820)

# train: SPRIOMICS
ridge.glmnet <- cv.glmnet(x = as.matrix(spiro.fev1[, -1]),
                          y = spiro.fev1$POST_FEV1_DERV_V1,
                          type.measure = "mse",
                          alpha = 0)
ridge.glmnet$lambda.min

best_ridge_coef <- as.numeric(coef(ridge.glmnet, s = ridge.glmnet$lambda.min))[-1]

adalasso.spiro.cv <- cv.glmnet(x = as.matrix(spiro.fev1[, -1]),
                          y = spiro.fev1$POST_FEV1_DERV_V1,
                          type.measure = "mse",
                          alpha = 1, 
                          penalty.factor = 1/abs(best_ridge_coef))

plot(adalasso.spiro.cv)

best_adalasso_spiro_coef <- coef(adalasso.spiro.cv, s = adalasso.spiro.cv$lambda.min, exact=TRUE)
```

## Table 2 & S1: All Metabolites selected by Adaptive LASSO
```{r trainSpiro_metabs}
#Capture Metabs used in FEV1 Adaptive LASSO
alasso_est <- best_adalasso_spiro_coef[, "s1"][-1]

fev1.adalasso.metab <- best_adalasso_spiro_coef[, "s1"][best_adalasso_spiro_coef[, "s1"] != 0]
fev1.adalasso.metab <- fev1.adalasso.metab[-1]

fev1.metabs <- as.data.frame(cbind(names(fev1.adalasso.metab), fev1.adalasso.metab))
names(fev1.metabs) <- c("metab_id", "ALASSO_FEV1_Metabs")
fev1.metabs <- merge(fev1.metabs, var.key, by = "metab_id", all.x = T)
# write.csv(fev1.metabs, file = file.path(metRSfolder, "FEV1_ALASSO_Metabolites_trainSPIRO.csv"),
#           row.names = F)
```

```{r trainSpiro_regr}
### SPIROMICS linear Models
adalasso.spiro.pred <- predict(adalasso.spiro.cv, as.matrix(spiro.fev1[, -1]), 
                               s = adalasso.spiro.cv$lambda.min)
adalasso.spiro.pred <- as.data.frame(adalasso.spiro.pred)
names(adalasso.spiro.pred) <- "adalasso_pred"
adalasso.spiro.pred$SUBJID <- rownames(adalasso.spiro.pred)

spiro.c$adalasso_pred <- NULL
adalasso.spiro.pred <- merge(spiro.c, adalasso.spiro.pred, by = "SUBJID")

spiro.adalasso <- lm(POST_FEV1_DERV_V1 ~ adalasso_pred, data = adalasso.spiro.pred)
spiro.adalasso.r2 <- round(summary(spiro.adalasso)$adj.r.squared*100, 3) 
spiro.adalasso.mse <- fxn.mse(spiro.adalasso$model$POST_FEV1_DERV_V1, fitted(spiro.adalasso))


spiro.adalasso.adj <- lm(POST_FEV1_DERV_V1 ~ adalasso_pred + GENDER + AGE_DERV_V1 + 
                      RACE.ETH+ BMI_CM_V1 + smokStatus + SMOKING_PACK_YEARS_V1 + SITE_cat,
                      data = adalasso.spiro.pred)

spiro.adalasso.adj.r2 <- round(summary(spiro.adalasso.adj)$adj.r.squared*100, 3) 
spiro.adalasso.adj.mse <- fxn.mse(spiro.adalasso.adj$model$POST_FEV1_DERV_V1, fitted(spiro.adalasso.adj))

### COPDGene Linear Models
adalasso.copdg.pred <- predict(adalasso.spiro.cv, as.matrix(copdg.fev1[, -1]), 
                               s = adalasso.spiro.cv$lambda.min)

adalasso.copdg.pred <- as.data.frame(adalasso.copdg.pred)
names(adalasso.copdg.pred) <- "adalasso_pred"
adalasso.copdg.pred$sid <- rownames(adalasso.copdg.pred)

copd <- merge(copd.clin.p2, adalasso.copdg.pred, by = "sid")

adalasso.copdg <- lm(FEV1_post.2 ~ adalasso_pred, data = copd)
copdg.adalasso.r2 <-round(summary(adalasso.copdg)$adj.r.squared*100, 3) 
copdg.adalasso.mse <- fxn.mse(adalasso.copdg$model$FEV1_post.2, fitted(adalasso.copdg))


adalasso.copdg.adj <- lm(FEV1_post.2 ~ adalasso_pred + gender + age_visit + race.eth +
                           BMI + smoking_status + ATS_PackYears + ccenter, 
                    data = copd)

copdg.adalasso.adj.r2 <- round(summary(adalasso.copdg.adj)$adj.r.squared*100, 3)
copdg.adalasso.adj.mse <-fxn.mse(adalasso.copdg.adj$model$FEV1_post.2, fitted(adalasso.copdg.adj)) 

fev1.adaLASSO.tS$metRS_Only_R2 <- c(spiro.adalasso.r2, copdg.adalasso.r2)
fev1.adaLASSO.tS$metRS_Only_MSE <- c(spiro.adalasso.mse, copdg.adalasso.mse)

fev1.adaLASSO.tS$Cov_metRS_R2 <- c(spiro.adalasso.adj.r2, copdg.adalasso.adj.r2)
fev1.adaLASSO.tS$Cov_metRS_MSE <- c(spiro.adalasso.adj.mse, copdg.adalasso.adj.mse)

fev1.adaLASSO.tS$met_num <- nrow(fev1.metabs)

fev1.adaLASSO.tS <- fev1.adaLASSO.tS[, c("study", "met_num", 
                                   "Cov_Only_R2", "metRS_Only_R2", "Cov_metRS_R2",
                                   "Cov_Only_MSE", "metRS_Only_MSE", "Cov_metRS_MSE")]

# write.csv(fev1.adaLASSO.tS, file.path(metRSfolder, "FEV1_adaLASSO_trainSpiro.csv"),
#           row.names = F)
```

```{r alasso_plots}
# jpeg(filename = file.path(metRSfolder, "SPIROMICS_FEV1_predvsobs_regLine.jpg"),
#      height = 5, width = 5, units = "in", res = 300)
plot(adalasso.spiro.pred$adalasso_pred, adalasso.spiro.pred$POST_FEV1_DERV_V1,
     xlim = c(min(adalasso.spiro.pred$POST_FEV1_DERV_V1, 
                  adalasso.spiro.pred$adalasso_pred, na.rm = T), 
              max(adalasso.spiro.pred$POST_FEV1_DERV_V1, 
                  adalasso.spiro.pred$adalasso_pred, na.rm = T)), 
     ylim = c(min(adalasso.spiro.pred$POST_FEV1_DERV_V1, 
                  adalasso.spiro.pred$adalasso_pred, na.rm = T), 
              max(adalasso.spiro.pred$POST_FEV1_DERV_V1, 
                  adalasso.spiro.pred$adalasso_pred, na.rm = T)),
     xlab = "Predicted FEV1 -- Adaptive LASSO", 
     ylab = "Observed FEV1", 
     pch = 16, cex = 0.5)
# ####Color points by self-reported smoking status
# points(adalasso.spiro.pred$adalasso_pred[adalasso.spiro.pred$smokStatus == "Current Smoker"], 
#        adalasso.spiro.pred$POST_FEV1_DERV_V1[adalasso.spiro.pred$smokStatus == "Current Smoker"], 
#        col = "pink", pch = 16)
# points(adalasso.spiro.pred$adalasso_pred[adalasso.spiro.pred$smokStatus == "Former Smoker"], 
#        adalasso.spiro.pred$POST_FEV1_DERV_V1[adalasso.spiro.pred$smokStatus == "Former Smoker"], 
#        col = "violet", pch = 16)
# points(adalasso.spiro.pred$adalasso_pred[adalasso.spiro.pred$smokStatus == "Never Smoked"], 
#        adalasso.spiro.pred$POST_FEV1_DERV_V1[adalasso.spiro.pred$smokStatus == "Never Smoked"], 
#        col = "lightblue", pch = 16)
abline(0,1, col = "blue")
abline(lm(POST_FEV1_DERV_V1 ~ adalasso_pred, data = adalasso.spiro.pred), col="darkorange")
# dev.off()

plot(adalasso.spiro.pred$POST_FEV1_DERV_V1, adalasso.spiro.pred$POST_FEV1_DERV_V1 - adalasso.spiro.pred$adalasso_pred,
     xlab = "Observed FEV1", 
     ylab = "Difference Obs - Predicted")
abline(0,0, col = "blue")


## COPDGene plots

# jpeg(filename = file.path(metRSfolder, "COPDGene_FEV1_predvsobs_regLine.jpg"),
#      height = 5, width = 5, units = "in", res = 300)
plot(copd$adalasso_pred, copd$FEV1_post.2,
     xlim = c(min(copd$FEV1_post.2, copd$adalasso_pred, na.rm = T), 
              max(copd$FEV1_post.2, copd$adalasso_pred, na.rm = T)), 
     ylim = c(min(copd$FEV1_post.2, copd$adalasso_pred, na.rm = T), 
              max(copd$FEV1_post.2, copd$adalasso_pred, na.rm = T)),
     xlab = "Predicted FEV1 -- Adaptive LASSO", 
     ylab = "Observed FEV1", 
     pch = 16, cex = 0.5)
# ####Color points by self-reported smoking status
# points(copd$adalasso_pred[copd$smoking_status == "Former Smoker"],
#        copd$FEV1_post.2[copd$smoking_status == "Former Smoker"],
#        col = "violet", pch = 16)
# points(copd$adalasso_pred[copd$smoking_status == "Current Smoker"],
#        copd$FEV1_post.2[copd$smoking_status == "Current Smoker"],
#        col = "pink", pch = 16)
# points(copd$adalasso_pred[copd$smoking_status == "Never Smoked"],
#        copd$FEV1_post.2[copd$smoking_status == "Never Smoked"],
#        col = "lightblue", pch = 16)
abline(0,1, col = "blue")
abline(lm(FEV1_post.2 ~ adalasso_pred, data = copd), col="darkorange")

# dev.off()

plot(copd$FEV1_post.2, copd$FEV1_post.2 - copd$adalasso_pred,
     xlab = "Observed FEV1", 
     ylab = "Difference Obs - Predicted")
abline(0,0, col = "blue")
```

## Sensitivity Analysis -- 
### Train on COPDGene, Validation on SPIROMICS
```{r train_copdg}
set.seed(20210820)
# train: COPDGene
ridge.glmnet <- cv.glmnet(x = as.matrix(copdg.fev1[, -1]),
                          y = copdg.fev1$FEV1_post.2,
                          type.measure = "mse",
                          alpha = 0)
ridge.glmnet$lambda.min

best_ridge_coef <- as.numeric(coef(ridge.glmnet, s = ridge.glmnet$lambda.min))[-1]

adalasso.spiro <- glmnet(x = as.matrix(copdg.fev1[, -1]),
                         y = copdg.fev1$FEV1_post.2,
                         alpha = 1,
                         penalty.factor = 1/abs(best_ridge_coef))

adalasso.spiro.cv <- cv.glmnet(x = as.matrix(copdg.fev1[, -1]),
                          y = copdg.fev1$FEV1_post.2,
                          type.measure = "mse",
                          alpha = 1,
                          penalty.factor = 1/abs(best_ridge_coef))

plot(adalasso.spiro.cv)

best_adalasso_spiro_coef <- coef(adalasso.spiro.cv, s = adalasso.spiro.cv$lambda.min)
```

```{r trainCopdg_metabs}
#Capture Metabs used in FEV1 Adaptive LASSO
alasso_est <- best_adalasso_spiro_coef[, "s1"][-1]

fev1.adalasso.metab <- best_adalasso_spiro_coef[, "s1"][best_adalasso_spiro_coef[, "s1"] != 0]
fev1.adalasso.metab <- fev1.adalasso.metab[-1]

fev1.metabs <- as.data.frame(cbind(names(fev1.adalasso.metab), fev1.adalasso.metab))
names(fev1.metabs) <- c("metab_id", "ALASSO_FEV1_Metabs")
fev1.metabs <- merge(fev1.metabs, var.key, by = "metab_id", all.x = T)
# write.csv(fev1.metabs, file = file.path(metRSfolder, "FEV1_ALASSO_Metabolites_trainCOPDGene.csv"),
#           row.names = F)
```

```{r trainCopdg_regr}
### SPIROMICS linear Models
adalasso.spiro.pred <- predict(adalasso.spiro.cv, as.matrix(spiro.fev1[, -1]), 
                               s = adalasso.spiro.cv$lambda.min)
adalasso.spiro.pred <- as.data.frame(adalasso.spiro.pred)
names(adalasso.spiro.pred) <- "adalasso_pred"
adalasso.spiro.pred$SUBJID <- rownames(adalasso.spiro.pred)

spiro.c$adalasso_pred <- NULL
adalasso.spiro.pred <- merge(spiro.c, adalasso.spiro.pred, by = "SUBJID")

spiro.adalasso <- lm(POST_FEV1_DERV_V1 ~ adalasso_pred, data = adalasso.spiro.pred)
spiro.adalasso.r2 <- round(summary(spiro.adalasso)$adj.r.squared*100, 3) 
spiro.adalasso.mse <- fxn.mse(spiro.adalasso$model$POST_FEV1_DERV_V1, fitted(spiro.adalasso))


spiro.adalasso.adj <- lm(POST_FEV1_DERV_V1 ~ adalasso_pred + GENDER + AGE_DERV_V1 + 
                      RACE.ETH+ BMI_CM_V1 + smokStatus + SMOKING_PACK_YEARS_V1 + SITE_cat,
                      data = adalasso.spiro.pred)

spiro.adalasso.adj.r2 <- round(summary(spiro.adalasso.adj)$adj.r.squared*100, 3) 
spiro.adalasso.adj.mse <- fxn.mse(spiro.adalasso.adj$model$POST_FEV1_DERV_V1, fitted(spiro.adalasso.adj))

### COPDGene Linear Models
adalasso.copdg.pred <- predict(adalasso.spiro.cv, as.matrix(copdg.fev1[, -1]), 
                               s = adalasso.spiro.cv$lambda.min)

adalasso.copdg.pred <- as.data.frame(adalasso.copdg.pred)
names(adalasso.copdg.pred) <- "adalasso_pred"
adalasso.copdg.pred$sid <- rownames(adalasso.copdg.pred)

copd <- merge(copd.clin.p2, adalasso.copdg.pred, by = "sid")

adalasso.copdg <- lm(FEV1_post.2 ~ adalasso_pred, data = copd)
copdg.adalasso.r2 <-round(summary(adalasso.copdg)$adj.r.squared*100, 3) 
copdg.adalasso.mse <- fxn.mse(adalasso.copdg$model$FEV1_post.2, fitted(adalasso.copdg))


adalasso.copdg.adj <- lm(FEV1_post.2 ~ adalasso_pred + gender + age_visit + race.eth +
                           BMI + smoking_status + ATS_PackYears + ccenter, 
                    data = copd)

copdg.adalasso.adj.r2 <- round(summary(adalasso.copdg.adj)$adj.r.squared*100, 3)
copdg.adalasso.adj.mse <-fxn.mse(adalasso.copdg.adj$model$FEV1_post.2, fitted(adalasso.copdg.adj)) 

fev1.adaLASSO.tC$metRS_Only_R2 <- c(spiro.adalasso.r2, copdg.adalasso.r2)
fev1.adaLASSO.tC$metRS_Only_MSE <- c(spiro.adalasso.mse, copdg.adalasso.mse)

fev1.adaLASSO.tC$Cov_metRS_R2 <- c(spiro.adalasso.adj.r2, copdg.adalasso.adj.r2)
fev1.adaLASSO.tC$Cov_metRS_MSE <- c(spiro.adalasso.adj.mse, copdg.adalasso.adj.mse)

fev1.adaLASSO.tC$met_num <- nrow(fev1.metabs)

fev1.adaLASSO.tC <- fev1.adaLASSO.tC[, c("study", "met_num", 
                                   "Cov_Only_R2", "metRS_Only_R2", "Cov_metRS_R2",
                                   "Cov_Only_MSE", "metRS_Only_MSE", "Cov_metRS_MSE")]

# write.csv(fev1.adaLASSO.tC, file.path(metRSfolder, "FEV1_adaLASSO_trainCOPDGene.csv"),
#           row.names = F)
```

```{r trainCopdg_alasso_plots}
# jpeg(filename = file.path(metRSfolder, "SPIROMICS_FEV1_predvsobs_regLine.jpg"),
#      height = 5, width = 5, units = "in", res = 300)
plot(adalasso.spiro.pred$adalasso_pred, adalasso.spiro.pred$POST_FEV1_DERV_V1,
     xlim = c(min(adalasso.spiro.pred$POST_FEV1_DERV_V1, 
                  adalasso.spiro.pred$adalasso_pred, na.rm = T), 
              max(adalasso.spiro.pred$POST_FEV1_DERV_V1, 
                  adalasso.spiro.pred$adalasso_pred, na.rm = T)), 
     ylim = c(min(adalasso.spiro.pred$POST_FEV1_DERV_V1, 
                  adalasso.spiro.pred$adalasso_pred, na.rm = T), 
              max(adalasso.spiro.pred$POST_FEV1_DERV_V1, 
                  adalasso.spiro.pred$adalasso_pred, na.rm = T)),
     xlab = "Predicted FEV1 -- Adaptive LASSO", 
     ylab = "Observed FEV1", 
     pch = 16, cex = 0.5)
# ####Color points by self-reported smoking status
# points(adalasso.spiro.pred$adalasso_pred[adalasso.spiro.pred$smokStatus == "Current Smoker"], 
#        adalasso.spiro.pred$POST_FEV1_DERV_V1[adalasso.spiro.pred$smokStatus == "Current Smoker"], 
#        col = "pink", pch = 16)
# points(adalasso.spiro.pred$adalasso_pred[adalasso.spiro.pred$smokStatus == "Former Smoker"], 
#        adalasso.spiro.pred$POST_FEV1_DERV_V1[adalasso.spiro.pred$smokStatus == "Former Smoker"], 
#        col = "violet", pch = 16)
# points(adalasso.spiro.pred$adalasso_pred[adalasso.spiro.pred$smokStatus == "Never Smoked"], 
#        adalasso.spiro.pred$POST_FEV1_DERV_V1[adalasso.spiro.pred$smokStatus == "Never Smoked"], 
#        col = "lightblue", pch = 16)
abline(0,1, col = "blue")
abline(lm(POST_FEV1_DERV_V1 ~ adalasso_pred, data = adalasso.spiro.pred), col="darkorange")
# dev.off()

plot(adalasso.spiro.pred$POST_FEV1_DERV_V1, adalasso.spiro.pred$POST_FEV1_DERV_V1 - adalasso.spiro.pred$adalasso_pred,
     xlab = "Observed FEV1", 
     ylab = "Difference Obs - Predicted")
abline(0,0, col = "blue")


## COPDGene plots

# jpeg(filename = file.path(metRSfolder, "COPDGene_FEV1_predvsobs_regLine.jpg"),
#      height = 5, width = 5, units = "in", res = 300)
plot(copd$adalasso_pred, copd$FEV1_post.2,
     xlim = c(min(copd$FEV1_post.2, copd$adalasso_pred, na.rm = T), 
              max(copd$FEV1_post.2, copd$adalasso_pred, na.rm = T)), 
     ylim = c(min(copd$FEV1_post.2, copd$adalasso_pred, na.rm = T), 
              max(copd$FEV1_post.2, copd$adalasso_pred, na.rm = T)),
     xlab = "Predicted FEV1 -- Adaptive LASSO", 
     ylab = "Observed FEV1", 
     pch = 16, cex = 0.5)
# ####Color points by self-reported smoking status
# points(copd$adalasso_pred[copd$smoking_status == "Former Smoker"],
#        copd$FEV1_post.2[copd$smoking_status == "Former Smoker"],
#        col = "violet", pch = 16)
# points(copd$adalasso_pred[copd$smoking_status == "Current Smoker"],
#        copd$FEV1_post.2[copd$smoking_status == "Current Smoker"],
#        col = "pink", pch = 16)
# points(copd$adalasso_pred[copd$smoking_status == "Never Smoked"],
#        copd$FEV1_post.2[copd$smoking_status == "Never Smoked"],
#        col = "lightblue", pch = 16)
abline(0,1, col = "blue")
abline(lm(FEV1_post.2 ~ adalasso_pred, data = copd), col="darkorange")

# dev.off()

plot(copd$FEV1_post.2, copd$FEV1_post.2 - copd$adalasso_pred,
     xlab = "Observed FEV1", 
     ylab = "Difference Obs - Predicted")
abline(0,0, col = "blue")
```

### Train SPIROMICS with weights on highest and lowest quintile
```{r adalasso_spiroWT}
quan.fev1 <- quantile(spiro.fev1$POST_FEV1_DERV_V1, probs = c(0.2, 0.8))
wt <- ifelse(spiro.fev1$POST_FEV1_DERV_V1 < quan.fev1[1] | 
               spiro.fev1$POST_FEV1_DERV_V1 > quan.fev1[2], 
             5, 1)


set.seed(20210820)

# train: SPRIOMICS
ridge.glmnet <- cv.glmnet(x = as.matrix(spiro.fev1[, -1]),
                          y = spiro.fev1$POST_FEV1_DERV_V1,
                          type.measure = "mse",
                          alpha = 0, 
                          weights = wt)
ridge.glmnet$lambda.min

best_ridge_coef <- as.numeric(coef(ridge.glmnet, s = ridge.glmnet$lambda.min))[-1]

adalasso.spiro.cv <- cv.glmnet(x = as.matrix(spiro.fev1[, -1]),
                          y = spiro.fev1$POST_FEV1_DERV_V1,
                          type.measure = "mse",
                          alpha = 1, 
                          penalty.factor = 1/abs(best_ridge_coef), 
                          weights = wt)

plot(adalasso.spiro.cv)

best_adalasso_spiro_coef <- coef(adalasso.spiro.cv, s = adalasso.spiro.cv$lambda.min, exact=TRUE)
```

```{r trainSpiroWT_metabs}
#Capture Metabs used in FEV1 Adaptive LASSO
alasso_est <- best_adalasso_spiro_coef[, "s1"][-1]

fev1.adalasso.metab <- best_adalasso_spiro_coef[, "s1"][best_adalasso_spiro_coef[, "s1"] != 0]
fev1.adalasso.metab <- fev1.adalasso.metab[-1]

fev1.metabs <- as.data.frame(cbind(names(fev1.adalasso.metab), fev1.adalasso.metab))
names(fev1.metabs) <- c("metab_id", "ALASSO_FEV1_Metabs")
fev1.metabs <- merge(fev1.metabs, var.key, by = "metab_id", all.x = T)
# write.csv(fev1.metabs, file = file.path(metRSfolder, "FEV1_ALASSO_Metabolites_trainSPIRO_Weighted.csv"),
#           row.names = F)
```

```{r trainSpiroWT_regr}
### SPIROMICS linear Models
adalasso.spiro.pred <- predict(adalasso.spiro.cv, as.matrix(spiro.fev1[, -1]), 
                               s = adalasso.spiro.cv$lambda.min)
adalasso.spiro.pred <- as.data.frame(adalasso.spiro.pred)
names(adalasso.spiro.pred) <- "adalasso_pred"
adalasso.spiro.pred$SUBJID <- rownames(adalasso.spiro.pred)

spiro.c$adalasso_pred <- NULL
adalasso.spiro.pred <- merge(spiro.c, adalasso.spiro.pred, by = "SUBJID")

spiro.adalasso <- lm(POST_FEV1_DERV_V1 ~ adalasso_pred, data = adalasso.spiro.pred)
spiro.adalasso.r2 <- round(summary(spiro.adalasso)$adj.r.squared*100, 3) 
spiro.adalasso.mse <- fxn.mse(spiro.adalasso$model$POST_FEV1_DERV_V1, fitted(spiro.adalasso))


spiro.adalasso.adj <- lm(POST_FEV1_DERV_V1 ~ adalasso_pred + GENDER + AGE_DERV_V1 + 
                      RACE.ETH+ BMI_CM_V1 + smokStatus + SMOKING_PACK_YEARS_V1 + SITE_cat,
                      data = adalasso.spiro.pred)

spiro.adalasso.adj.r2 <- round(summary(spiro.adalasso.adj)$adj.r.squared*100, 3) 
spiro.adalasso.adj.mse <- fxn.mse(spiro.adalasso.adj$model$POST_FEV1_DERV_V1, fitted(spiro.adalasso.adj))

### COPDGene Linear Models
adalasso.copdg.pred <- predict(adalasso.spiro.cv, as.matrix(copdg.fev1[, -1]), 
                               s = adalasso.spiro.cv$lambda.min)

adalasso.copdg.pred <- as.data.frame(adalasso.copdg.pred)
names(adalasso.copdg.pred) <- "adalasso_pred"
adalasso.copdg.pred$sid <- rownames(adalasso.copdg.pred)

copd <- merge(copd.clin.p2, adalasso.copdg.pred, by = "sid")

adalasso.copdg <- lm(FEV1_post.2 ~ adalasso_pred, data = copd)
copdg.adalasso.r2 <-round(summary(adalasso.copdg)$adj.r.squared*100, 3) 
copdg.adalasso.mse <- fxn.mse(adalasso.copdg$model$FEV1_post.2, fitted(adalasso.copdg))


adalasso.copdg.adj <- lm(FEV1_post.2 ~ adalasso_pred + gender + age_visit + race.eth +
                           BMI + smoking_status + ATS_PackYears + ccenter, 
                    data = copd)

copdg.adalasso.adj.r2 <- round(summary(adalasso.copdg.adj)$adj.r.squared*100, 3)
copdg.adalasso.adj.mse <-fxn.mse(adalasso.copdg.adj$model$FEV1_post.2, fitted(adalasso.copdg.adj)) 

fev1.adaLASSO.tSwt$metRS_Only_R2 <- c(spiro.adalasso.r2, copdg.adalasso.r2)
fev1.adaLASSO.tSwt$metRS_Only_MSE <- c(spiro.adalasso.mse, copdg.adalasso.mse)

fev1.adaLASSO.tSwt$Cov_metRS_R2 <- c(spiro.adalasso.adj.r2, copdg.adalasso.adj.r2)
fev1.adaLASSO.tSwt$Cov_metRS_MSE <- c(spiro.adalasso.adj.mse, copdg.adalasso.adj.mse)

fev1.adaLASSO.tSwt$met_num <- nrow(fev1.metabs)

fev1.adaLASSO.tSwt <- fev1.adaLASSO.tSwt[, c("study", "met_num", 
                                   "Cov_Only_R2", "metRS_Only_R2", "Cov_metRS_R2",
                                   "Cov_Only_MSE", "metRS_Only_MSE", "Cov_metRS_MSE")]

# write.csv(fev1.adaLASSO.tSwt, file.path(metRSfolder, "FEV1_adaLASSO_trainSpiroWeighted.csv"),
#            row.names = F)
```

## Scatterplot of top 4 metabolites and FEV1
```{r scatter_topMetab}

fev1.metabs <- read.csv(file.path(metRSfolder, "FEV1_ALASSO_Metabolites_trainSPIRO.csv"), 
                        header = T, stringsAsFactors = F)
fev1.metabs <- fev1.metabs[order(abs(as.numeric(fev1.metabs$ALASSO_FEV1_Metabs)), decreasing = T), ]

top.metab <- head(fev1.metabs$metab_id, n = 4)

spiro.plots <- vector(mode="list", length =4)
copdg.plots <- vector(mode="list", length =4) 

ggplot(data = spiro.fev1, aes(spiro.fev1[, top.metab[1]], POST_FEV1_DERV_V1))+
    geom_point()+
    theme_bw()+
    theme(axis.title.x = element_blank(), 
          axis.title.y = element_blank())

ggplot(data = spiro.fev1, aes(spiro.fev1[, top.metab[2]], POST_FEV1_DERV_V1))+
    geom_point()+
    theme_bw()+
    theme(axis.title.x = element_blank(), 
          axis.title.y = element_blank())

for(i in seq_along(top.metab)){
  message(i)
  min.val <- min(spiro.fev1[top.metab[i]], copdg.fev1[, top.metab[i]])
  max.val <- max(spiro.fev1[top.metab[i]], copdg.fev1[, top.metab[i]])
  spiro.plots[[i]] <-  local({
    i <- i
    p1 <- ggplot(data = spiro.fev1, 
                              aes(x = spiro.fev1[, top.metab[i]], 
                                  y = POST_FEV1_DERV_V1))+
    geom_point()+
    xlim(min.val, max.val)+
    theme_bw()+
    theme(axis.title.x = element_blank(), 
          axis.title.y = element_blank())
    print(p1)
  })
    

  copdg.plots[[i]] <- local({
    i <- i
    p1 <- ggplot(data = copdg.fev1, 
                             aes(copdg.fev1[, top.metab[i]], 
                                 FEV1_post.2))+
      geom_point()+
      xlim(min.val, max.val)+
      theme_bw()+
      theme(axis.title.x = element_blank(), 
            axis.title.y = element_blank())
  })
    
}


ylab = textGrob("FEV1, post-bronchodilator (L)", rot = 90)
xlab1 = textGrob(fev1.metabs[fev1.metabs$metab_id == top.metab[1], "CHEMICAL_NAME"])
xlab2 = textGrob(fev1.metabs[fev1.metabs$metab_id == top.metab[2], "CHEMICAL_NAME"])
xlab3 = textGrob(fev1.metabs[fev1.metabs$metab_id == top.metab[3], "CHEMICAL_NAME"])
xlab4 = textGrob(fev1.metabs[fev1.metabs$metab_id == top.metab[4], "CHEMICAL_NAME"])

g1 <- grid.arrange(arrangeGrob(spiro.plots[[1]], top = textGrob("SPIROMICS")), 
                   arrangeGrob(copdg.plots[[1]], top = textGrob("COPDGene")), 
                  bottom = xlab1, ncol = 2)
g2 <- grid.arrange(spiro.plots[[2]], copdg.plots[[2]], 
                  bottom = xlab2, ncol = 2)
g3 <- grid.arrange(spiro.plots[[3]], copdg.plots[[3]], 
                  bottom = xlab3, ncol = 2)
g4 <- grid.arrange(spiro.plots[[4]], copdg.plots[[4]], 
                  bottom = xlab4, ncol = 2)

jpeg(filename = file.path(metRSfolder, "Scatterplot_topAdaLASSO_metabs_FEV1.jpg"), 
     width = 9, height = 11, res = 300, units = "in")
grid.arrange(g1, g2, g3, g4,
             left = ylab, 
             ncol = 1)
dev.off()
```

## Venn Diagram of metabolite overlap from switching training and validation sets
```{r overlapMetab}
fev1.trainSPIRO <- read.csv(file.path(metRSfolder, "FEV1_ALASSO_Metabolites_trainSPIRO.csv"), 
                            header = T, stringsAsFactors = F)
fev1.trainSPIRO$absCoeff <- abs(fev1.trainSPIRO$ALASSO_FEV1_Metabs)

metab.tS <- fev1.trainSPIRO$metab_id

fev1.trainCOPDG <- read.csv(file.path(metRSfolder, "FEV1_ALASSO_Metabolites_trainCOPDGene.csv"), 
                            header = T, stringsAsFactors = F)
fev1.trainCOPDG$absCoeff <- abs(fev1.trainCOPDG$ALASSO_FEV1_Metabs)

metab.tC <- fev1.trainCOPDG$metab_id

venn.list <- list("Train SPIROMICS" = metab.tS, 
                  "Train COPDGene" = metab.tC)

venn.plot <- venn.diagram(venn.list,
                          file.path(metRSfolder, "AdaptiveLASSO_Metabolites_SwitchTraining.tiff"))

inALL <- intersect(metab.tS, metab.tC)

View(fev1.trainSPIRO[fev1.trainSPIRO$metab_id %in% inALL, ])
View(fev1.trainCOPDG[fev1.trainCOPDG$metab_id %in% inALL, ])

```
